BotForge/requirements.txt
discord.py>=2.5.2
fastapi>=0.115.0
uvicorn[standard]>=0.31.0
aiosqlite>=0.20.0
PyNaCl>=1.5.0
